Updated definitive dataset, FuseWref and train
Definitive dataset generation is now seeded so the partitions can be reproduced any where. It now adds the name of the folder from the which the image was read as a new column and the images are divided into the folders they come from too.
FuseWRef was refactored to improve readability and support generic expressions and image search
Train now creates folders and images with an extra string defined by the user, this is to improve version control when training models with different parameters. It also supports pixel wise data-model inference